+++ src/tunnel/detail/platform_unsupported/layer_tun.hpp
@@ -18,13 +18,6 @@ template <class Super>
 class layer_tun : public Super
 {
 public:
-    void create(const std::string& interface_name, std::error_code& error)
-    {
-        (void)interface_name;
-
-        Super::do_log(log_level::error, log_kind::unsupported_platform);
-        error = std::make_error_code(std::errc::not_supported);
-    }
     void create(const std::string& interface_name, std::error_code& error,
                 bool vnet_hdr)
     {
@@ -33,6 +26,7 @@ public:
         Super::do_log(log_level::error, log_kind::unsupported_platform);
         error = std::make_error_code(std::errc::not_supported);
     }
+
     void rename(const std::string& interface_name, std::error_code& error) const
     {
         (void)interface_name;